requirements.txt
Pillow~=8.2.0
scipy~=1.7.0
numpy~=1.21.0
matplotlib~=3.4.2
scikit-image~=0.18.2